Rename *_t.cpp to *_t.hpp
